+++ weather_alerts/__init__.py
@@ -1,3 +1,3 @@
 """WeatherTelegramAlerts package."""
 
-__version__ = "2.1.4"
+__version__ = "2.2.5"